Set theme styles to be only created once and move them to a separate folder
1. Use `document.head` to replace
   `document.getElementsByTagName('head')[0]`
2. In previous version, if users call `applyTheme()` many times, it will
   generate the same css class many times with different contents.
   Therefore, split it to 2 parts. One is creating css class in
   constructor, and the other is changing the content in `applyTheme()`
3. Extract all these theme related methods to a separate module.
