+++ src/index.ts
@@ -1,9 +1,10 @@
 import * as styles from './styles.scss';
 import { PreferencesControl } from './preferencesControl';
 import { HtmlTools } from './htmlTools';
+import { ThemesController } from './themes/themesController';
 
 import { RTL_LANGUAGE } from './language-list.const';
-import { DEFAULT_THEMES } from './theme-styles';
+import { DEFAULT_THEMES } from './themes/theme-styles';
 
 import { ICookieCategory } from './interfaces/CookieCategories';
 import { ITextResources, IOptions, IThemes, ITheme } from './interfaces/Options';
@@ -99,6 +100,7 @@ export class ConsentControl {
             }
         }
 
+        ThemesController.createThemeStyle();
         if (options?.initialTheme) {
             this.applyTheme(options.initialTheme);
         }
@@ -129,138 +131,13 @@ export class ConsentControl {
      */
     public createTheme(name: string, theme: ITheme): void {
         let typeName = <keyof IThemes> name;
-
         this.themes[typeName] = theme;
-        let currentTheme = <ITheme> this.themes[typeName];
 
-        // Styles that can be determined by another one
-
-        // determined by "dialog-background-color"
-        if (!theme["background-color-between-page-and-dialog"]) {
-            let provided: string = theme["dialog-background-color"];
-            this.setMissingColorFromAnotherProperty("background-color-between-page-and-dialog", provided, 0.6, currentTheme);
-        }
-        if (!theme["primary-button-text-color"]) {
-            currentTheme["primary-button-text-color"] = theme["dialog-background-color"];
-        }
-        if (!theme["primary-button-disabled-text-color"]) {
-            currentTheme["primary-button-disabled-text-color"] = theme["dialog-background-color"];
-        }
-
-        // determined by "primary-button-color"
-        if (!theme["dialog-border-color"]) {
-            currentTheme["dialog-border-color"] = theme["primary-button-color"];
-        }
-        if (!theme["hyperlink-font-color"]) {
-            currentTheme["hyperlink-font-color"] = theme["primary-button-color"];
-        }
-        if (!theme["primary-button-hover-color"]) {
-            currentTheme["primary-button-hover-color"] = theme["primary-button-color"];
-        }
-        if (!theme["primary-button-disabled-color"]) {
-            currentTheme["primary-button-disabled-color"] = theme["primary-button-color"];
-        }
-        if (!theme["primary-button-border"]) {
-            currentTheme["primary-button-border"] = "1px solid " + theme["primary-button-color"];
-        }
-        if (!theme["primary-button-focus-border-color"]) {
-            currentTheme["primary-button-focus-border-color"] = theme["primary-button-color"];
-        }
-        if (!theme["radio-button-hover-border-color"]) {
-            currentTheme["radio-button-hover-border-color"] = theme["primary-button-color"];
-        }
-
-        // determined by "text-color"
-        if (!theme["secondary-button-text-color"]) {
-            currentTheme["secondary-button-text-color"] = theme["text-color"];
-        }
-        if (!theme["secondary-button-disabled-text-color"]) {
-            currentTheme["secondary-button-disabled-text-color"] = theme["text-color"];
-        }
-        if (!theme["radio-button-border-color"]) {
-            currentTheme["radio-button-border-color"] = theme["text-color"];
-        }
-        if (!theme["radio-button-checked-background-color"]) {
-            currentTheme["radio-button-checked-background-color"] = theme["text-color"];
-        }
-        if (!theme["radio-button-hover-background-color"]) {
-            let provided: string = theme["text-color"];
-            this.setMissingColorFromAnotherProperty("radio-button-hover-background-color", provided, 0.8, currentTheme);
-        }
-        if (!theme["radio-button-disabled-color"]) {
-            let provided: string = theme["text-color"];
-            this.setMissingColorFromAnotherProperty("radio-button-disabled-color", provided, 0.2, currentTheme);
-        }
-        if (!theme["radio-button-disabled-border-color"]) {
-            let provided: string = theme["text-color"];
-            this.setMissingColorFromAnotherProperty("radio-button-disabled-border-color", provided, 0.2, currentTheme);
-        }
-
-        // determined by "secondary-button-color"
-        if (!theme["secondary-button-hover-color"]) {
-            currentTheme["secondary-button-hover-color"] = theme["secondary-button-color"];
-        }
-
-        // determined by "secondary-button-disabled-color"
-        if (!theme["secondary-button-disabled-border"]) {
-            currentTheme["secondary-button-disabled-border"] = "1px solid " + theme["secondary-button-disabled-color"];
-        }
-        
-        // determined by "secondary-button-border"
-        if (!theme["secondary-button-hover-border"]) {
-            currentTheme["secondary-button-hover-border"] = theme["secondary-button-border"];
-        }
-        if (!theme["secondary-button-focus-border-color"]) {
-            let secondaryBtnBorderElement: string[] = theme["secondary-button-border"].split(" ");
-            currentTheme["secondary-button-focus-border-color"] = secondaryBtnBorderElement[secondaryBtnBorderElement.length - 1];
-        }
+        // Determine optional styles
+        ThemesController.createTheme(<ITheme> this.themes[typeName], theme);
     }
 
     /**
-     * Set missing color from another provided color
-     * 
-     * e.g.: 
-     * provider: #ffffff; missing: rgba(255, 255, 255, transparencyFactor)
-     * provider: rgb(255, 255, 255); missing: rgba(255, 255, 255, transparencyFactor)
-     * provider: rgb(255, 255, 255, 1); missing: rgba(255, 255, 255, transparencyFactor)
-     * 
-     * @param {keyof ITheme} missing missing property in the theme
-     * @param {string} provider the provided property in the theme
-     * @param {number} transparencyFactor the alpha channel that will be used in the missing property
-     * @param {ITheme} currentTheme the target theme that will be set
-     */
-    private setMissingColorFromAnotherProperty(missing: keyof ITheme, 
-                                               provider: string, 
-                                               transparencyFactor: number,  
-                                               currentTheme: ITheme): void {
-        
-        // provider: #ffffff; missing: rgba(255, 255, 255, transparencyFactor)
-        if (provider.startsWith("#")) {
-            let hexColor = parseInt(provider, 16);
-            let r = (hexColor >> 16) & 255;
-            let g = (hexColor >> 8) & 255;
-            let b = hexColor & 255;
-
-            currentTheme[missing] = "rgba(" + r + ", " + g + ", " + b + ", " + transparencyFactor + ")";
-        }
-        // provider: rgb(255, 255, 255); missing: rgba(255, 255, 255, transparencyFactor)
-        else if (provider.startsWith("rgb(")) {
-            let missingColor = "rgba" + provider.substring(3, provider.length - 1) + ", " + transparencyFactor + ")";
-            currentTheme[missing] = missingColor;
-        }
-        // provider: rgb(255, 255, 255, 1); missing: rgba(255, 255, 255, transparencyFactor)
-        else if (provider.startsWith("rgba")) {
-            let rgbIdx = provider.lastIndexOf(",");
-            let transparency = provider.substring(rgbIdx + 1, provider.length - 1);
-
-            let newTransparency = parseInt(transparency) * transparencyFactor;
-            let missingColor = provider.substring(0, rgbIdx + 1) + newTransparency + ")";
-            currentTheme[missing] = missingColor;
-        }
-    }
-
-    /**
-     * 
      * Apply the theme and change banner and preferences dialog's color 
      * 
      * @param {string} themeName theme that will be applied
@@ -271,226 +148,7 @@ export class ConsentControl {
         }
 
         let theme: ITheme = <ITheme> this.themes[themeName];
-
-        // overwrite banner styles
-        let bannerStyle = document.createElement('style');
-        bannerStyle.type = 'text/css';
-        bannerStyle.innerHTML = `.${ styles.bannerBody } {
-            background-color: ${ theme["banner-background-color"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(bannerStyle);
-
-        // overwrite text color styles
-        let textColorStyle = document.createElement('style');
-        textColorStyle.type = 'text/css';
-        textColorStyle.innerHTML = `.${ styles.textColorTheme } {
-            color: ${ theme["text-color"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(textColorStyle);
-        
-        // overwrite hyper link styles
-        let hyperLinkStyle = document.createElement('style');
-        hyperLinkStyle.type = 'text/css';
-        hyperLinkStyle.innerHTML = `.${ styles.hyperLinkTheme } a {
-            color: ${ theme["hyperlink-font-color"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(hyperLinkStyle);
-        
-        this.setDialogStyle(theme);
-        this.setPrimaryBtnStyle(theme);
-        this.setSecondaryBtnStyle(theme);
-        this.setRadioBtnStyle(theme);
-    }
-
-    /**
-     * Apply the theme to dialog
-     * 
-     * @param {ITheme} theme theme that will be applied to dialog
-     */
-    private setDialogStyle(theme: ITheme): void {
-
-        // overwrite cookieModal styles (background between page and dialog)
-        let cookieModalStyle = document.createElement('style');
-        cookieModalStyle.type = 'text/css';
-        cookieModalStyle.innerHTML = `.${ styles.cookieModal } {
-            background-color: ${ theme["background-color-between-page-and-dialog"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(cookieModalStyle);
-
-        // overwrite preferences dialog styles
-        let dialogStyle = document.createElement('style');
-        dialogStyle.type = 'text/css';
-        dialogStyle.innerHTML = `.${ styles.dialogTheme } {
-            background-color: ${ theme["dialog-background-color"] } !important;
-            border: 1px solid ${ theme["dialog-border-color"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(dialogStyle);
-        
-        // overwrite close button styles
-        let closeIconStyle = document.createElement('style');
-        closeIconStyle.type = 'text/css';
-        closeIconStyle.innerHTML = `.${ styles.closeModalIcon } {
-            color: ${ theme["close-button-color"] } !important;
-            background-color: ${ theme["dialog-background-color"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(closeIconStyle);
-
-    }
-
-    /**
-     * Apply the theme to secondary button
-     * 
-     * @param {ITheme} theme theme that will be applied to secondary button
-     */
-    private setSecondaryBtnStyle(theme: ITheme): void {
-        let secondaryButtonStyle = document.createElement('style');
-        secondaryButtonStyle.type = 'text/css';
-        secondaryButtonStyle.innerHTML = `.${ styles.secondaryButtonTheme } { 
-            border: ${ theme["secondary-button-border"] } !important;
-            background-color: ${ theme["secondary-button-color"] } !important;
-            color: ${ theme["secondary-button-text-color"] } !important;
-        }`;
-
-        let secondaryButtonHoverStyle = document.createElement('style');
-        secondaryButtonHoverStyle.type = 'text/css';
-        secondaryButtonHoverStyle.innerHTML = `.${ styles.secondaryButtonTheme }:hover {
-            color: ${ theme["secondary-button-text-color"] } !important;
-            background-color: ${ theme["secondary-button-hover-color"] } !important;
-            box-shadow: ${ theme["secondary-button-hover-shadow"] } !important;
-            border: ${ theme["secondary-button-hover-border"] } !important;
-        }`;
-
-        let secondaryButtonFocusStyle = document.createElement('style');
-        secondaryButtonFocusStyle.type = 'text/css';
-        secondaryButtonFocusStyle.innerHTML = `.${ styles.secondaryButtonTheme }:focus {
-            background-color: ${ theme["secondary-button-hover-color"] } !important;
-            box-shadow: ${ theme["secondary-button-hover-shadow"] } !important;
-            border: 2px solid ${ theme["secondary-button-focus-border-color"] } !important;
-        }`;
-
-        let secondaryButtonDisabledStyle = document.createElement('style');
-        secondaryButtonDisabledStyle.type = 'text/css';
-        secondaryButtonDisabledStyle.innerHTML = `.${ styles.secondaryButtonTheme }:disabled {
-            opacity: ${ theme["secondary-button-disabled-opacity"] } !important;
-            color: ${ theme["secondary-button-disabled-text-color"] } !important;
-            background-color: ${ theme["secondary-button-disabled-color"] } !important;
-            border: ${ theme["secondary-button-disabled-border"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(secondaryButtonStyle);
-        document.getElementsByTagName('head')[0].appendChild(secondaryButtonHoverStyle);
-        document.getElementsByTagName('head')[0].appendChild(secondaryButtonFocusStyle);
-        document.getElementsByTagName('head')[0].appendChild(secondaryButtonDisabledStyle);
-    }
-
-    /**
-     * Apply the theme to primary button
-     * 
-     * @param {ITheme} theme theme that will be applied to primary button
-     */
-    private setPrimaryBtnStyle(theme: ITheme): void {
-        let primaryButtonStyle = document.createElement('style');
-        primaryButtonStyle.type = 'text/css';
-        primaryButtonStyle.innerHTML = `.${ styles.primaryButtonTheme } { 
-            border: ${ theme["primary-button-border"] } !important;
-            background-color: ${ theme["primary-button-color"] } !important;
-            color: ${ theme["primary-button-text-color"] } !important;
-        }`;
-
-        let primaryButtonHoverStyle = document.createElement('style');
-        primaryButtonHoverStyle.type = 'text/css';
-        primaryButtonHoverStyle.innerHTML = `.${ styles.primaryButtonTheme }:hover {
-            color: ${ theme["primary-button-text-color"] } !important;
-            background-color: ${ theme["primary-button-hover-color"] } !important;
-            box-shadow: ${ theme["primary-button-hover-shadow"] } !important;
-            border: ${ theme["primary-button-hover-border"] } !important;
-        }`;
-
-        let primaryButtonFocusStyle = document.createElement('style');
-        primaryButtonFocusStyle.type = 'text/css';
-        primaryButtonFocusStyle.innerHTML = `.${ styles.primaryButtonTheme }:focus {
-            background-color: ${ theme["primary-button-hover-color"] } !important;
-            box-shadow: ${ theme["primary-button-hover-shadow"] } !important;
-            border: 2px solid ${ theme["primary-button-focus-border-color"] } !important;
-        }`;
-
-        let primaryButtonDisabledStyle = document.createElement('style');
-        primaryButtonDisabledStyle.type = 'text/css';
-        primaryButtonDisabledStyle.innerHTML = `.${ styles.primaryButtonTheme }:disabled {
-            opacity: ${ theme["primary-button-disabled-opacity"] } !important;
-            color: ${ theme["primary-button-disabled-text-color"] } !important;
-            background-color: ${ theme["primary-button-disabled-color"] } !important;
-            border: ${ theme["primary-button-disabled-border"] } !important;
-        }`;
-
-        document.getElementsByTagName('head')[0].appendChild(primaryButtonStyle);
-        document.getElementsByTagName('head')[0].appendChild(primaryButtonHoverStyle);
-        document.getElementsByTagName('head')[0].appendChild(primaryButtonFocusStyle);
-        document.getElementsByTagName('head')[0].appendChild(primaryButtonDisabledStyle);
-    }
-
-    /**
-     * Apply the theme to radio button
-     * 
-     * @param {ITheme} theme theme that will be applied to radio button
-     */
-    private setRadioBtnStyle(theme: ITheme): void {
-        let radioButtonStyle = document.createElement('style');
-        radioButtonStyle.type = 'text/css';
-        radioButtonStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn } + span::before {
-            border: 1px solid ${ theme["radio-button-border-color"] } !important;
-            background-color: ${ theme["dialog-background-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonStyle);
-
-        let radioButtonCheckedStyle = document.createElement('style');
-        radioButtonCheckedStyle.type = 'text/css';
-        radioButtonCheckedStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn }:checked + span::after {
-            background-color: ${ theme["radio-button-checked-background-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonCheckedStyle);
-
-        let radioButtonHoverStyle = document.createElement('style');
-        radioButtonHoverStyle.type = 'text/css';
-        radioButtonHoverStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn } + span:hover::before {
-            border: 1px solid ${ theme["radio-button-hover-border-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonHoverStyle);
-
-        let radioButtonHoverAfterStyle = document.createElement('style');
-        radioButtonHoverAfterStyle.type = 'text/css';
-        radioButtonHoverAfterStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn } + span:hover::after {
-            background-color: ${ theme["radio-button-hover-background-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonHoverAfterStyle);
-
-        let radioButtonFocusStyle = document.createElement('style');
-        radioButtonFocusStyle.type = 'text/css';
-        radioButtonFocusStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn } + span:focus::before {
-            border: 1px solid ${ theme["radio-button-hover-border-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonFocusStyle);
-
-        let radioButtonFocusAfterStyle = document.createElement('style');
-        radioButtonFocusAfterStyle.type = 'text/css';
-        radioButtonFocusAfterStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn } + span:focus::after {
-            background-color: ${ theme["radio-button-checked-background-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonFocusAfterStyle);
-
-        let radioButtonDisabledStyle = document.createElement('style');
-        radioButtonDisabledStyle.type = 'text/css';
-        radioButtonDisabledStyle.innerHTML = `input[type="radio"].${ styles.cookieItemRadioBtn }:disabled + span::before {
-            border: 1px solid ${ theme["radio-button-disabled-border-color"] } !important;
-            background-color: ${ theme["radio-button-disabled-color"] } !important;
-        }`;
-        document.getElementsByTagName('head')[0].appendChild(radioButtonDisabledStyle);
+        ThemesController.applyTheme(theme);
     }
 
     /**
@@ -506,7 +164,7 @@ export class ConsentControl {
             let meta = document.createElement('meta');
             meta.name = "viewport";
             meta.content = "width=device-width, initial-scale=1.0";
-            document.getElementsByTagName('head')[0].appendChild(meta);
+            document.head.appendChild(meta);
         }
 
         // Remove existing banner and preference dialog
